requirements.txt
streamlit>=1.35.0
llama-index-core>=0.10.0
llama-index-agent-openai>=0.2.0
llama-index-llms-openai>=0.1.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
plotly>=5.17.0
